# Split common and extended sets of likely subtags