Part of https://github.com/unicode-org/icu4x/issues/2903

See https://unicode-org.atlassian.net/browse/CLDR-15071

See https://unicode-org.atlassian.net/browse/CLDR-16438