@sffc 2023-02-27
@macchiati Here's my first attempt at splitting the likelysubtags into a "common" and "extended" set. I filtered them based on whether the language subtag of the maximized locale is present in coverageLevels.json at "basic" level or higher.

This brings likely subtags data from 19K to 5K against CLDR 42 data (not yet including the new stuff in CLDR 43).

Is this a reasonable set?

CC @zbraniecki