@sffc 2023-02-27
Bench results:

Before my changes:

<details>
<pre>
uncanonicalized/clone   time:   [2.9712 µs 2.9786 µs 2.9853 µs]                                   
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
uncanonicalized/canonicalize                                                                             
                        time:   [33.334 µs 33.369 µs 33.407 µs]
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

canonicalized/clone     time:   [1.1913 µs 1.1917 µs 1.1920 µs]                                 
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe
canonicalized/canonicalize                                                                             
                        time:   [2.0785 µs 2.0804 µs 2.0824 µs]
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe

likelysubtags/maximize  time:   [1.7754 µs 1.7775 µs 1.7799 µs]                                    
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe
</pre>
</details>

After https://github.com/unicode-org/icu4x/pull/3148:

<details>
<pre>
uncanonicalized/clone   time:   [2.9410 µs 2.9435 µs 2.9471 µs]                                   
                        change: [-0.7867% -0.4780% -0.1866%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 7 outliers among 100 measurements (7.00%)
  1 (1.00%) low mild
  3 (3.00%) high mild
  3 (3.00%) high severe
uncanonicalized/canonicalize                                                                             
                        time:   [33.847 µs 33.859 µs 33.871 µs]
                        change: [+1.2963% +1.4037% +1.5067%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) low mild
  3 (3.00%) high mild
  1 (1.00%) high severe

canonicalized/clone     time:   [1.2048 µs 1.2059 µs 1.2071 µs]                                 
                        change: [+1.0260% +1.1696% +1.3424%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 8 outliers among 100 measurements (8.00%)
  4 (4.00%) high mild
  4 (4.00%) high severe
canonicalized/canonicalize                                                                             
                        time:   [2.1296 µs 2.1332 µs 2.1366 µs]
                        change: [+2.3027% +2.4648% +2.6224%] (p = 0.00 < 0.05)
                        Performance has regressed.

likelysubtags/maximize  time:   [1.7620 µs 1.7650 µs 1.7691 µs]                                    
                        change: [-0.6843% -0.4982% -0.2884%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high severe
</pre>
</details>

With this PR:

<details>
<pre>
uncanonicalized/clone   time:   [2.9605 µs 2.9639 µs 2.9682 µs]                                   
                        change: [+0.5486% +0.6796% +0.8215%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) low mild
  3 (3.00%) high severe
uncanonicalized/canonicalize                                                                             
                        time:   [34.167 µs 34.199 µs 34.247 µs]
                        change: [+0.8556% +0.9317% +1.0064%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) low mild
  1 (1.00%) high mild
  1 (1.00%) high severe

canonicalized/clone     time:   [1.2119 µs 1.2126 µs 1.2133 µs]                                 
                        change: [+0.2987% +0.4915% +0.7000%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild
  2 (2.00%) high severe
canonicalized/canonicalize                                                                             
                        time:   [2.0721 µs 2.0757 µs 2.0801 µs]
                        change: [-2.9349% -2.7676% -2.5965%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe

likelysubtags/maximize  time:   [1.8958 µs 1.8979 µs 1.9003 µs]                                    
                        change: [+7.0578% +7.2696% +7.4618%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild
</pre>
</details>