@srl295 2023-02-27
i don't know what your size goals are specifically

but you might consider including languages which are some % of the spoken population even if not basic or not even in cldr.  because this is supplemental, not related to locale maturity in CLDR.

the languagepop tables have info on which langs are even 'on the radar' in a territory.

see `territoryInfo.json`

```json
      "AD": {
        "_gdp": "3327000000",
        "_literacyPercent": "100",
        "_population": "77000",
        "languagePopulation": {
          "ca": {
            "_populationPercent": "51",
            "_officialStatus": "official"
          },
          "es": {
            "_populationPercent": "43"
          },
          "fr": {
            "_populationPercent": "7.5"
          }
        }
      },
```