@macchiati 2023-03-08
This looks like a reasonable set. @srl295 I think it's ok to restrict to the supported locales for small-footprint devices. The number of situations where one would need to canonicalize a locale that is not supported is probably pretty small.